AMPOPS-38 Merging and Branch Related Tasks AMP 1.16.2 -> AMP 1.16.6 merged into AMP 2.0
git-svn-id: file:///svn/branches/AMP_2_0_RELEASE@16041 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/dbentity/AmpActivityProgram.java
@@ -9,7 +9,7 @@ import org.digijava.module.aim.util.ProgramUtil;
 public class AmpActivityProgram implements Versionable, Serializable, Cloneable {
 
         private Long ampActivityProgramId;
-        private Long programPercentage;
+        private Float programPercentage;
         private AmpTheme program;
         private AmpActivityVersion activity;
         private AmpActivityProgramSettings programSetting;
@@ -17,7 +17,7 @@ public class AmpActivityProgram implements Versionable, Serializable, Cloneable 
                 return ampActivityProgramId;
         }
 
-        public Long getProgramPercentage() {
+        public Float getProgramPercentage() {
                 return programPercentage;
         }
 
@@ -37,7 +37,7 @@ public class AmpActivityProgram implements Versionable, Serializable, Cloneable 
                 this.ampActivityProgramId = ampActivityProgramId;
         }
 
-        public void setProgramPercentage(Long programPercentage) {
+        public void setProgramPercentage(Float programPercentage) {
                 this.programPercentage = programPercentage;
         }
